Move AnchorRenderer
