+++ src/components/HTMLEngineProvider/HTMLRenderers/htmlRendererPropTypes.js
@@ -1,0 +1,5 @@
+import PropTypes from 'prop-types';
+
+export default {
+    tnode: PropTypes.object,
+};
